==> scripts/8_AMOVA_pairwiseFST.qmd <==
---
title: "AMOVA"
format: html
editor: visual
---

```{r}
library(here)
library(tidyverse)
library(poppr)
library(vcfR)
library(StAMPP)
```

```{r}
acro_vcf<- read.vcfR(here("data/vcf/acro_filt.recode.vcf"))
acro_gl = vcfR2genlight(acro_vcf, n.cores = 10) 
```

```{r}
dat_meta_acro <- read.csv(here("data/meta/acro_filt_samples_clean.csv"))
strata(acro_gl) = data.frame(dat_meta_acro)
ploidy(acro_gl) <- 2
```

```{r}
set.seed(1999)
poppr.amova(acro_gl, ~short) %>% 
  randtest(nrepet = 99) #no
```

```{r}
poppr.amova(acro_gl, ~size) %>% 
  randtest(nrepet = 99) #no
```

```{r}
poppr.amova(acro_gl, ~site) %>% 
  randtest(nrepet = 99) #no
```

```{r}
poppr.amova(acro_gl, ~depth) %>% 
  randtest(nrepet = 99) #almost
```

```{r}
setPop(acro_gl) = ~short
fst_short <- stamppFst(acro_gl, 
                      nboots = 99, 
                      percent = 95, 
                      nclusters = 8) 
fst_short$Fsts %>% round(4)
fst_short$Pvalues %>% round(4)
```

```{r}
setPop(acro_gl) = ~size
fst_size <- stamppFst(acro_gl, 
                      nboots = 99, 
                      percent = 95, 
                      nclusters = 8) 
fst_size$Fsts %>% round(4)
fst_size$Pvalues %>% round(4)
```

```{r}
setPop(acro_gl) = ~site
fst_site <- stamppFst(acro_gl, 
                      nboots = 99, 
                      percent = 95, 
                      nclusters = 8) 
fst_site$Fsts %>% round(4)
fst_site$Pvalues %>% round(4)
```

```{r}
setPop(acro_gl) = ~depth
fst_depth <- stamppFst(acro_gl, 
                      nboots = 99, 
                      percent = 95, 
                      nclusters = 8) 
fst_depth$Fsts %>% round(4)
fst_depth$Pvalues %>% round(4)#almost
```
